Refactor bofire utilities
