+++ bark/bofire_utils/data_models/api.py
@@ -1,5 +1,10 @@
-from bark.bofire_utils.data_models.strategies import TreeKernelStrategy  # noqa: F401
-from bark.bofire_utils.data_models.surrogates import (  # noqa: F401
+from bark.bofire_utils.data_models.strategies.tree_kernel import (
+    TreeKernelStrategy,  # noqa: F401
+)
+from bark.bofire_utils.data_models.surrogates.bark import (  # noqa: F401
+    BARKPriorSurrogate,
     BARKSurrogate,
-    LeafGPSurrogate,
+)
+from bark.bofire_utils.data_models.surrogates.leafgp import (
+    LeafGPSurrogate,  # noqa: F401
 )
