+++ bark/bofire_utils/data_models/mapper.py
@@ -5,12 +5,14 @@ from bofire.surrogates.api import map as bofire_map_surrogate
 from bofire.surrogates.surrogate import Surrogate
 
 from bark.bofire_utils.data_models import api as bark_data_models
-from bark.bofire_utils.strategies import TreeKernelStrategy
-from bark.bofire_utils.surrogates import BARKSurrogate, LeafGPSurrogate
+from bark.bofire_utils.strategies.tree_kernel import TreeKernelStrategy
+from bark.bofire_utils.surrogates.bark import BARKPriorSurrogate, BARKSurrogate
+from bark.bofire_utils.surrogates.leafgp import LeafGPSurrogate
 
 SURROGATE_MAP: dict[type[surrogates_data_models.Surrogate], type[Surrogate]] = {
     bark_data_models.LeafGPSurrogate: LeafGPSurrogate,
     bark_data_models.BARKSurrogate: BARKSurrogate,
+    bark_data_models.BARKPriorSurrogate: BARKPriorSurrogate,
 }
 
 STRATEGY_MAP = {
