+++ bark/bofire_utils/data_models/surrogates/leafgp.py
@@ -1,0 +1,6 @@
+from bofire.data_models.surrogates.surrogate import Surrogate
+from bofire.data_models.surrogates.trainable import TrainableSurrogate
+
+
+class LeafGPSurrogate(Surrogate, TrainableSurrogate):
+    pass
