+++ bark/bofire_utils/strategies/tree_kernel.py
@@ -8,10 +8,12 @@ from bofire.data_models.types import InputTransformSpecs
 from bofire.strategies.api import RandomStrategy
 from bofire.strategies.predictives.predictive import PredictiveStrategy
 
-from bark.bofire_utils.data_models.strategies import TreeKernelStrategy as DataModel
+from bark.bofire_utils.data_models.strategies.tree_kernel import (
+    TreeKernelStrategy as DataModel,
+)
 
 # from bark.bofire_utils.data_models.mapper import surrogate_map
-from bark.bofire_utils.surrogates import BARKSurrogate
+from bark.bofire_utils.surrogates.bark import BARKSurrogate
 from bark.optimizer.opt_core import get_opt_core_from_domain
 from bark.optimizer.opt_model import build_opt_model_from_forest
 from bark.optimizer.proposals import propose
